requirements.txt
# Framework principal
Flask==3.0.2
Flask-SQLAlchemy==3.1.1

# Banco de dados
psycopg2-binary==2.9.9
SQLAlchemy==2.0.28

# Configuração e variáveis de ambiente
python-decouple==3.8

# Serialização/Deserialização
marshmallow==3.21.1

# Documentação API (Swagger)
flask-restx==1.3.0

# CORS (caso precise)
Flask-CORS==4.0.0

# Gerenciamento de datas
python-dateutil==2.9.0

# Validações
email-validator==2.1.1